Replace old-style NIPSA worker with Celery tasks
This replaces the old NSQ-driven mailer worker with two distinct Celery
tasks -- one for removing and one for adding the NIPSA flag to a user's
annotations.

In order to simplify the testing of these tasks, I've split the
responsibilities in `h.nipsa.worker` up into `bulk_update_annotations`,
which accepts a query and an "action function" which is used to provide
update actions for each annotation in the query set, and the Celery
tasks `add_nipsa` and `remove_nipsa`, which merely pass appropriate
parameters to `bulk_update_annotations`.

I've also fixed what would appear to be a bug in the original worker
implementation, where the query was executed against all open indices,
rather than only against the index configured in the application
settings.