Remove old-style notification worker
The old "notification" worker listened to the all-annotation event
stream from NSQ, and for each annotation potentially generated one or
more notification emails and sent them.

This commit moves the generation of the emails to be sent into the web
process, in a subscriber for `h.api.events.AnnotationEvent`. This
removes some boilerplate for creating an Annotation object from
serialized data, and also ensures that the generation of those
emails occurs within the request transaction.

Instead of having a custom worker, we then dispatch the mails to be sent
using the existing `h.mailer` Celery task.

N.B. Since the `h.notification` package is known to be buggy, the body
of the event handler is wrapped in a try/except so that errors in
generation notification emails don't cause the annotation CRUD action to
be aborted. This is a temporary measure -- when we've fixed the known
bugs in `generate_notifications` this can be removed.